refactor(hub-common): move private helpers into an _internal folder
affects: @esri/hub-common
