+++ packages/common/src/downloads/_internal/canCreateExportItem.ts
