+++ packages/common/src/downloads/_internal/file-url-fetchers/fetchHubApiDownloadFileUrl.ts
@@ -6,7 +6,7 @@ import {
   IFetchDownloadFileUrlOptions,
   ServiceDownloadFormat,
   downloadProgressCallback,
-} from "../types";
+} from "../../types";
 
 /**
  * @private
