+++ packages/common/src/downloads/_internal/format-fetchers/getExportImageFormats.ts
@@ -1,4 +1,4 @@
-import { IDynamicDownloadFormat } from "../types";
+import { IDynamicDownloadFormat } from "../../types";
 
 /**
  * @private
