+++ packages/common/src/downloads/_internal/isMapOrFeatureServiceEntity.ts
