+++ packages/common/src/downloads/_internal/isServiceEntity.ts
