+++ packages/common/src/downloads/index.ts
@@ -1,4 +1,4 @@
-export * from "./helpers/build-existing-exports-portal-query";
-export * from "./helpers/types";
+export * from "./build-existing-exports-portal-query";
+export * from "./types";
 export * from "./fetchDownloadFileUrl";
 export * from "./fetchDownloadFormats";
