+++ packages/common/src/downloads/types.ts
@@ -1,5 +1,5 @@
-import { IArcGISContext } from "../../ArcGISContext";
-import { IHubEditableContent } from "../../core/types/IHubEditableContent";
+import { IArcGISContext } from "../ArcGISContext";
+import { IHubEditableContent } from "../core/types/IHubEditableContent";
 
 /**
  * This hash map was defined to support the previous implementation of the export item flow.
@@ -85,68 +85,6 @@ export enum ServiceDownloadFormat {
 }
 
 /**
- * Formats supported by the /export endpoint of the Portal API.
- */
-export const EXPORT_ITEM_FORMATS = [
-  ServiceDownloadFormat.CSV,
-  ServiceDownloadFormat.KML,
-  ServiceDownloadFormat.SHAPEFILE,
-  ServiceDownloadFormat.FILE_GDB,
-  ServiceDownloadFormat.GEOJSON,
-  ServiceDownloadFormat.EXCEL,
-  ServiceDownloadFormat.FEATURE_COLLECTION,
-] as const;
-
-export type ExportItemFormat = (typeof EXPORT_ITEM_FORMATS)[number];
-
-/**
- * Formats supported by the /exportImage endpoint of Image Services.
- */
-export const EXPORT_IMAGE_FORMATS = [
-  ServiceDownloadFormat.BIP,
-  ServiceDownloadFormat.BMP,
-  ServiceDownloadFormat.BSQ,
-  ServiceDownloadFormat.GIF,
-  ServiceDownloadFormat.JPG,
-  ServiceDownloadFormat.JPG_PNG,
-  ServiceDownloadFormat.LERC,
-  ServiceDownloadFormat.PNG,
-  ServiceDownloadFormat.PNG24,
-  ServiceDownloadFormat.PNG32,
-  ServiceDownloadFormat.PNG8,
-  ServiceDownloadFormat.TIFF,
-] as const;
-export type ExportImageFormat = (typeof EXPORT_IMAGE_FORMATS)[number];
-
-/**
- * Formats supported by the paging operation endpoint of the Hub Download API.
- */
-export const HUB_PAGING_JOB_FORMATS = [
-  ServiceDownloadFormat.CSV,
-  ServiceDownloadFormat.GEOJSON,
-  ServiceDownloadFormat.KML,
-  ServiceDownloadFormat.SHAPEFILE,
-] as const;
-export type HubPagingJobFormat = (typeof HUB_PAGING_JOB_FORMATS)[number];
-
-/**
- * Known formats supported by the /createReplica endpoint of the Hub Download API.
- * NOTE: this is may be incomplete and should be updated as needed.
- */
-export const CREATE_REPLICA_FORMATS = [
-  ServiceDownloadFormat.CSV,
-  ServiceDownloadFormat.EXCEL,
-  ServiceDownloadFormat.FEATURE_COLLECTION,
-  ServiceDownloadFormat.FILE_GDB,
-  ServiceDownloadFormat.GEOJSON,
-  ServiceDownloadFormat.GEO_PACKAGE,
-  ServiceDownloadFormat.JSON,
-  ServiceDownloadFormat.SHAPEFILE,
-  ServiceDownloadFormat.SQLITE,
-] as const;
-export type CreateReplicaFormat = (typeof CREATE_REPLICA_FORMATS)[number];
-
-/**
  * Represents a file format related to a content entity that can be downloaded.
  * Formats can either be dynamic (i.e., generated on-the-fly) or static (i.e., pre-generated)
  */
@@ -220,7 +158,7 @@ export interface IFetchDownloadFormatsOptions {
 /**
  * Options for instantiating an ArcgisHubDownloadError object.
  */
-export interface IArcgisHubDownloadErrorOptions {
+interface IArcgisHubDownloadErrorOptions {
   rawMessage: string; // raw error message
   messageId?: string; // well-known error message ID
   operation?: string; // operation that the error occurred in
